fix  last ad second lock no playback support ad bitrate selection

+++ playkit/src/main/java/com/kaltura/playkit/plugins/ads/ima/IMAConfig.java
@@ -38,7 +38,7 @@ public class IMAConfig {
     private AdResponseType adResponseType;
     private boolean enableBackgroundPlayback;
     private boolean autoPlayAdBreaks;
-    private int videoBitrate;
+    private int videoBitrate; // in KB
     private boolean adAttribution;
     private boolean adCountDown;
     private int  adLoadTimeOut;
@@ -98,7 +98,7 @@ public class IMAConfig {
         return this;
     }
 
-    public long getVideoBitrate() {
+    public int getVideoBitrate() {
         return videoBitrate;
     }
